+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: shorwood <shorwood@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/07/09 22:08:58 by shorwood     #+#   ##    ##    #+#       */
-/*   Updated: 2018/10/12 08:19:13 by shorwood    ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/10/15 21:06:37 by shorwood    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -18,7 +18,7 @@ int		ft_strncmp(const char *str, const char *cmp, size_t len)
 	while ((*str || *cmp) && len--)
 	{
 		if (*str != *cmp)
-			return ((unsigned char)*str - (unsigned char)*cmp);
+			return (int)((unsigned char)*str - (unsigned char)*cmp);
 		str++;
 		cmp++;
 	}
